requirements.txt
streamlit==1.29.0
openai>=2.14.0
python-dotenv==1.0.0
pandas
